Hopefully fixed MultiLoc
The problem was that the locationList property of MultiLoc was being used
for two incompatible purposes: maintaining a current list of rooms and
Things in which the MultiLoc is currently located, and defining a list
of Rooms, Things and *Regions* in which the MultiLoc is meant to start
out. This fix introduces an initialLocationList property for the
second purpose, and copies the locationProperty to it if it looks like
the user has defined that instead.