Alternative implentation of Ground
This defines a Floor class much like Mark's DefaultGround, but
defines the behaviour on the class. It then defines a defaultGround
object of the Floor class, which by default is added to every
Room (although this can be customized). The vocab for the
defaultGround object is defined in english.t.