Fixed Doer.redirect reporting bug
The previous version of Doer.redirect() created a new action for
each item the command iterated over, thereby obliterating any
previous reportList, so that the report stage could never work
properly. Doer.redirect() now only creates a new action if the
action it wants to redirect to is different from the Command's
current action, and if it does create a new action it first
lets the previous one report on what it's done by calling
curCmd.action.reportAction.