Improve handling of transparent enclosing SubComponents
This changes the matchName logic on SubComponents such that when
their parent is in scope, matching on the parent's vocabulary
gets a -15 likelihood adjustment rather than not matching at all.
This way, doing something like TOUCH CAR from the inside the car
will match on the interior, rather than insisting on touching
the exterior and telling you 'You can't reach the car through
the car'.